// lib/types/interfaces.ts
export interface iSettings {
  containerIPs: string[];
  dockerhub?: any;
  manageDNS: boolean;
  existingVPC?: boolean;
  existingVPCID?: string;
  existingSubnetIDs: Array<{ ID: string; AZ: string; type: string }>;
  existingRDS?: boolean;
  existingRDSEndpoint?: string;
  newRDSConfig?: {
    username: string;
    instanceType: string;
    deletionProtection: boolean;
  };
  github: {
    oauthToken: string;
    connectionID: string;
  };
  serversAlwaysOn: boolean;
  startHour?: string;
  stopHour?: string;
  ECSConfig: {
    minCapacity: number;
    maxCapacity: number;
    desiredCapacity: number;
  };
  Route53ResolverConfig: {
    ips: Array<CfnResolverEndpoint.IpAddressRequestProperty>
  },
  domainName: string;
  sslCertificateId?: string;
  otherSecrets?: any[];
  sslCertificateArn?: string;
  msTeamsWebhook?: string;
  bucketName?: string;
  eksClusterName: string;
  eksClusterPipelineName: string;
}